requirements.txt
fastapi
uvicorn 
fastapi-mail 
python-dotenv 
pandas 
pymysql 
python-jose 
sqlalchemy
bcrypt==3.2.0
redis>=4.5.0
python-multipart 
cryptography
passlib
fastapi-admin==0.3.3
xhtml2pdf
aiomysql
setuptools 
httpx
itsdangerous>=2.1.2



# create virtual environment ->  python -m venv myenv       
# Start Virtual Environment ->   venv\Scripts\activate   
# docker start redis:>  docker run -d redis -p 6379:6379 redis
#         output:14f4f9e4701d1a536d01d46ffdbc5995a47f1876aebc524b   :> not Exact but similar
#
#               if not Showing Image then run this command
#                                         docker run -d -p 6379:6379 redis
#
#
#  check Your Docker server :>   docker ps
#      CONTAINER ID   IMAGE     COMMAND                  CREATED              STATUS              PORTS                    NAMES
#      14f4f9e4701d   redis     "docker-entrypoint.s…"   About a minute ago   Up About a minute   0.0.0.0:6379->6379/tcp   upbeat_hypatia
#                  Thats  means Your Docker is running Smoothly
#
#terminal Code :>     docker exec -it 14f4f9e4701d{Container Id()rename it} redis-cli ping
#                Output Will be:>  PONG

#       Redis is running Smoothly
# docker exec -it b0353ffb94a7 redis-cli ping
#  run the Backend server:> uvicorn main:app --reload